# openapi/purplship-openapi_2021.3.yml
openapi: 3.0.1
info:
  title: Purplship API
  description: "\n## API Reference\n\nPurplship is an open source multi-carrier shipping\
    \ API that simplifies the integration of logistic carrier services.\n\nThe Purplship\
    \ API is organized around REST. Our API has predictable resource-oriented URLs,\
    \ accepts JSON-encoded request\nbodies, returns JSON-encoded responses, and uses\
    \ standard HTTP response codes, authentication, and verbs.\n\nThe Purplship API\
    \ differs for every account as we release new versions.\nThese docs are customized\
    \ to your version of the API.\n\n\n## Versioning\n\nWhen backwards-incompatible\
    \ changes are made to the API, a new, dated version is released. \nThe current\
    \ version is `2021.2.1`. \n\nRead our API changelog and to learn more about backwards\
    \ compatibility.\n\nAs a precaution, use API versioning to check a new API version\
    \ before committing to an upgrade.\n"
  contact:
    email: hello@purplship.com
  version: 2021.2.1
servers:
- url: https://app.purplship.com/
security:
- Token: []
...
paths:
...
  /v1/parcels:
    get:
      tags:
      - Parcels
      summary: List all parcels
      description: Retrieve all stored parcels.
      operationId: $$$list
      parameters:
      - name: limit
        in: query
        description: Number of results to return per page.
        schema:
          type: integer
      - name: offset
        in: query
        description: The initial index from which to return the results.
        schema:
          type: integer
      responses:
        200:
          description: ""
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ParcelList'
        400:
          description: ""
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      tags:
      - Parcels
      summary: Create a parcel
      description: Create a new parcel.
      operationId: $$$create
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ParcelData'
        required: true
      responses:
        200:
          description: ""
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Parcel'
        400:
          description: ""
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      x-codegen-request-body-name: data
  /v1/parcels/{id}:
    get:
      tags:
      - Parcels
      summary: Retrieve a parcel
      description: Retrieve a parcel.
      operationId: $$$retrieve
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      responses:
        200:
          description: ""
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Parcel'
        400:
          description: ""
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      tags:
      - Parcels
      summary: Remove a parcel
      description: Remove a parcel.
      operationId: $$$discard
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      responses:
        200:
          description: ""
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Operation'
        400:
          description: ""
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    patch:
      tags:
      - Parcels
      summary: Update a parcel
      description: modify an existing parcel's details.
      operationId: $$$update
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ParcelData'
        required: true
      responses:
        200:
          description: ""
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Parcel'
        400:
          description: ""
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      x-codegen-request-body-name: data
...
components:
  schemas:
...
    Message:
      type: object
      properties:
        carrier_name:
          title: Carrier name
          minLength: 1
          type: string
          description: The targeted carrier
        carrier_id:
          title: Carrier id
          minLength: 1
          type: string
          description: The targeted carrier name (unique identifier)
        message:
          title: Message
          minLength: 1
          type: string
          description: The error or warning message
        code:
          title: Code
          minLength: 1
          type: string
          description: The message code
        details:
          title: Details
          type: object
          additionalProperties:
            type: string
            nullable: true
          description: any additional details
      description: The list of error messages
    ErrorResponse:
      type: object
      properties:
        messages:
          type: array
          description: The list of error messages
          items:
            $ref: '#/components/schemas/Message'
...
    Operation:
      required:
      - operation
      - success
      type: object
      properties:
        operation:
          title: Operation
          minLength: 1
          type: string
          description: Operation performed
        success:
          title: Success
          type: boolean
          description: Specify whether the operation was successful
    Parcel:
      required:
      - weight
      - weight_unit
      type: object
      properties:
        id:
          title: Id
          minLength: 1
          type: string
          description: A unique identifier
        weight:
          title: Weight
          type: number
          description: The parcel's weight
        width:
          title: Width
          type: number
          description: The parcel's width
          nullable: true
        height:
          title: Height
          type: number
          description: The parcel's height
          nullable: true
        length:
          title: Length
          type: number
          description: The parcel's length
          nullable: true
        packaging_type:
          title: Packaging type
          type: string
          description: |2

            The parcel's packaging type.

            **Note that the packaging is optional when using a package preset**

            values: <br/>- **envelope**<br/>- **pak**<br/>- **tube**<br/>- **pallet**<br/>- **small_box**<br/>- **medium_box**<br/>- **your_packaging**

            For specific carriers packaging type, please consult [the reference](#operation/references).
          nullable: true
        package_preset:
          title: Package preset
          type: string
          description: |2

            The parcel's package preset.

            For specific carriers package preset, please consult [the reference](#operation/references).
          nullable: true
        description:
          title: Description
          type: string
          description: The parcel's description
          nullable: true
        content:
          title: Content
          type: string
          description: The parcel's content description
          nullable: true
        is_document:
          title: Is document
          type: boolean
          description: Indicates if the parcel is composed of documents only
          nullable: true
          default: false
        weight_unit:
          title: Weight unit
          type: string
          description: The parcel's weight unit
          enum:
          - KG
          - LB
        dimension_unit:
          title: Dimension unit
          type: string
          description: The parcel's dimension unit
          nullable: true
          enum:
          - CM
          - IN
    ParcelList:
      required:
      - results
      type: object
      properties:
        next:
          title: Next
          minLength: 1
          type: string
          format: uri
        previous:
          title: Previous
          minLength: 1
          type: string
          format: uri
        results:
          type: array
          items:
            $ref: '#/components/schemas/Parcel'
    ParcelData:
      required:
      - weight
      - weight_unit
      type: object
      properties:
        weight:
          title: Weight
          type: number
          description: The parcel's weight
        width:
          title: Width
          type: number
          description: The parcel's width
          nullable: true
        height:
          title: Height
          type: number
          description: The parcel's height
          nullable: true
        length:
          title: Length
          type: number
          description: The parcel's length
          nullable: true
        packaging_type:
          title: Packaging type
          type: string
          description: |2

            The parcel's packaging type.

            **Note that the packaging is optional when using a package preset**

            values: <br/>- **envelope**<br/>- **pak**<br/>- **tube**<br/>- **pallet**<br/>- **small_box**<br/>- **medium_box**<br/>- **your_packaging**

            For specific carriers packaging type, please consult [the reference](#operation/references).
          nullable: true
        package_preset:
          title: Package preset
          type: string
          description: |2

            The parcel's package preset.

            For specific carriers package preset, please consult [the reference](#operation/references).
          nullable: true
        description:
          title: Description
          type: string
          description: The parcel's description
          nullable: true
        content:
          title: Content
          type: string
          description: The parcel's content description
          nullable: true
        is_document:
          title: Is document
          type: boolean
          description: Indicates if the parcel is composed of documents only
          nullable: true
          default: false
        weight_unit:
          title: Weight unit
          type: string
          description: The parcel's weight unit
          enum:
          - KG
          - LB
        dimension_unit:
          title: Dimension unit
          type: string
          description: The parcel's dimension unit
          nullable: true
          enum:
          - CM
          - IN
...
  securitySchemes:
    Token:
      type: apiKey
      description: "\nThe Purplship API uses API keys to authenticate requests. You\
        \ can view and manage your \nAPI keys in the Purplship Dashboard.\n\nYour\
        \ API keys carry many privileges, so be sure to keep them secure! Do not share\
        \ your secret \nAPI keys in publicly accessible areas such as GitHub, client-side\
        \ code, and so forth.\n\nAuthentication to the API is performed via HTTP Bearer\
        \ Auth. You do not need to provide a password.\nTo authenticate via bearer\
        \ auth (e.g., for a cross-origin request), \nuse `-H \"Authorization: Token\
        \ 19707922d97cef7a5d5e17c331ceeff66f226660\"`.\n\nAPI requests without authentication\
        \ will also fail.\n"
      name: Authorization
      in: header